Lab 2 (#9)
* removed .idea folder

removed auto-generated IDE dirs

* deque finished

passed all checks

Note:
This version uses post-increment implementation for both the head and the tail. It is initialized to 0 and upon the first insert, head and tail are both moved.

There are a few different ways of doing a copy in the resize method. One is simply pushing off the front while copying over to resized array. The other breaks it into two iterations head -> size - 1 and 0 to tail. Another is using memcpy to copy if unwrapped and just keeping head/tail at their current positions.

* Delete workspace.xml

deleted .idea folder

* Update .gitignore